factor out some test utilities
